feat : 인터페이스 -> 골격클래스로 리팩토링
