+++ src/main/java/nextstep/courses/domain/session/impl/FreeSession.java
@@ -1,31 +1,14 @@
 package nextstep.courses.domain.session.impl;
 
-import nextstep.courses.entity.BaseEntity;
 import nextstep.courses.domain.cover.Image;
 import nextstep.courses.domain.session.RegistrationCount;
-import nextstep.courses.domain.session.Session;
 import nextstep.courses.domain.session.SessionName;
 import nextstep.courses.domain.session.SessionStatus;
 import nextstep.courses.domain.session.ValidityPeriod;
 import nextstep.courses.entity.SessionEntity;
 import nextstep.payments.domain.Money;
-import nextstep.payments.domain.Payment;
 
-public class FreeSession extends BaseEntity implements Session {
-
-    private Long id;
-
-    private final SessionName sessionName;
-
-    private final RegistrationCount registrationCount;
-
-    private final Money tuitionFee;
-
-    private final Image image;
-
-    private final SessionStatus sessionStatus;
-
-    private final ValidityPeriod validityPeriod;
+public class FreeSession extends AbstractSession {
 
     public FreeSession(SessionName sessionName, Image image, SessionStatus sessionStatus,
         ValidityPeriod validityPeriod) {
@@ -51,62 +34,6 @@ public class FreeSession extends BaseEntity implements Session {
     public FreeSession(Long id, SessionName sessionName, RegistrationCount registrationCount,
         Money tuitionFee,
         Image image, SessionStatus sessionStatus, ValidityPeriod validityPeriod) {
-        this.id = id;
-        this.sessionName = sessionName;
-        this.registrationCount = registrationCount;
-        this.tuitionFee = tuitionFee;
-        this.image = image;
-        this.sessionStatus = sessionStatus;
-        this.validityPeriod = validityPeriod;
-    }
-
-    @Override
-    public void addRegistrationCount() {
-        registrationCount.addValue();
-    }
-
-    @Override
-    public boolean isRegistrationAvailable() {
-        return isRecruitmentOpen(sessionStatus);
-    }
-
-    @Override
-    public boolean isPaymentAmountSameTuitionFee(Payment payment) {
-        return payment.isSamePaymentAmount(tuitionFee);
-    }
-
-    @Override
-    public Long getId() {
-        return id;
-    }
-
-    @Override
-    public String getSessionName() {
-        return sessionName.getValue();
-    }
-
-    @Override
-    public int getRegistrationCount() {
-        return registrationCount.getValue();
-    }
-
-    @Override
-    public int getTuitionFee() {
-        return tuitionFee.getValue();
-    }
-
-    @Override
-    public Image getImage() {
-        return image;
-    }
-
-    @Override
-    public String getSessionStatus() {
-        return sessionStatus.getValue();
-    }
-
-    @Override
-    public ValidityPeriod getValidityPeriod() {
-        return validityPeriod;
+        super(id, sessionName, registrationCount, tuitionFee, image, sessionStatus, validityPeriod);
     }
 }
